cmd/flux-start: Add --caliper-profile option
Add --caliper-profile=PROFILE option to flux-start, which causes

 LD_PRELOAD=libcaliper.so

and

 CALI_CONFIG_PROFILE=$PROFILE

to be set in the environment before invoking brokers. If Flux was
not built with --enable-caliper, then this option issues an error.

Unless already set in environment, CALI_LOG_VERBOSITY=0 is also set
by default to quiet the Caliper logs to stderr.