libpmi: provide a subset of PMI-2 API externally
Problem: when OpenMPI is configured to use SLURM's pmi2 MPI
personality, it links against SLURM's libpmi2.so, which connects
to SLURM's PMI-2 server using the PMI-2 wire protocol.
Unfortunately, SLURM's implementation ignores the protocol version
handshake and does not negotiate down to version 1.1 to work
with Flux like MPICH does natively.

Solution: Emulate a subset of the PMI-2 API and provide it in
a Flux libpmi2.so.  Coerce OpenMPI programs to use it by setting
LD_LIBRARY_PATH.  This libpmi2.so is implemented in terms of the
PMI-1 API, which in turn invokes the Flux PMI-1.1 wire protocol.

Fixes #746