travis-ci: automate GitHub release upload on tags
Setup automated release uploading to GitHub releases via the Travis
`deploy` section. Currently, the releases are hard-coded to be a
prelease, the prelease boolean should be removed once non-prelease
versions of flux-core are being tagged.

There is an issue with creating multi-line release "body" (Release
Notes) via Travis, so for now, we place a link to the in-tree
NEWS.md file, with an anchor to the most recent entry.

A condition is set to ensure that we only attempt to deploy from
the first job of a Travis build, see `on.condition` in the deploy
section. This assumes that TRAVIS_JOB_NUMBER is constructed of the
BUILD number and a numeric job number starting at 1.