libpmi: conform client bufsize to server
Problem: simple_client and simple_server separately hardwired
the buffer size for wire protocol messages.  If the server's value
is larger than the client's, the client cannot handle key, value,
and name strings that are within limits obtained from the server
but too large to be marshalled by the client.

Drop the pmi_simple_server_get_maxrequest() server API accessor
and define SIMPLE_MAX_PROTO_LINE in simple_server.h.  This is an
internal header - no reason to hide the fact that this value is
static.  It allows a per-client malloc to be dropped in the
flux-start(1) PMI server code and the pmi_simple.t unit test.

Instead of a static buffer, the client goes through init using
SIMPLE_MAX_PROTO_LINE, then calculates a new buffer size based
on the "maxes" returned in the handshake and afterwards uses that.