libpmi: debug_trace callback to server
In the simple_server, tracing was accomplished by directly
calling fprintf (stderr, ...) to dump request/response messages.
Replace this with a callback, and implement callbacks in
flux-start(1) and wrexecd server instances.  Besides allowing
for more flexible disposition of trace data in the future,
this allows the rank for a message to be determined and added
to he output (the rank is not known inside the server implementation).