test/wreck: ls test allows out of order listing
flux-wreck ls sometimes returns results out of numerical
jobid order.  Sort the output numerically before checking
that the test job appears last in the list.

Note: this was failing intermittently when the KVS commit
interval was increased, which may have affected timing of
KVS state updates for job run in the previous test.