broker/hello: use reduction for wireup scalability
Problem: "hello" wireup protocol doesn't scale.

Instead of having each rank send a message directly to rank 0
to show that it has wired up, implement a reduction handle so
messages are collected at each node of the TBON and a sum is
passed to rank 0.

The computed topology is used to establish a HWM value at each node
of the TBON.  The "tbon.descendants" broker attribute was previously
set up in anticipation of this need.  If all goes well, wireup
completes in lock step, with each level of the TBON reporting upstream
once its downstream nodes have reported in.  The computed HWM
value may be overridden by setting the hello.hwm broker attribute
on the broker command line.  setting it to zero disables the HWM
based reduction.

In addition, a timeout is set on the reduction handle so that after
a scaled fraction of the timeout based on the TBON level, the HWM
value is ignored and any messages are sent upstream without delay.
The default timeout (10s) can be manipulated by setting the
hello.timeout broker attribute on the broker command line.
Setting it to zero disables the timeout.

Note that the broker -T,--timeout option was removed and had a
different meaning; it would cause the wireup to abort once the
timeout was exceeded.  If thus functionality is required at some
point it should be re-added as a broker attribute.

Progress at rank 0 is logged at LOG_INFO, at each reduction sink
operation.

Before wireup starts, PMI exchanges endpoint information through
the enclosing instance.  Although there is a PMI_Barrier() between
the PMI write and read phases, one was needed after the read phase
so that PMI and wireup times reported on rank 0 properly reflect
instance-wide PMI and wireup times; otherwise wireup might start
at different times on different ranks.