libflux/reduce: drop internal timer scaling
Problem: timeout scaling is unnecessarily complicated and internally
use flux_attr_get() to obtain TBON geometry, which doesn't currently
work in broker-resident services.

Drop the internal scaling.  The purpose of the scaling was to maximize
reduction benefit for a particular monitoring use case, where batches
of items to be reduced were generated simultaneously across the instance.
In that case, the timeout can be calculated externally and passed to
flux_reduce_create() or flux_reduce_opt_set().  In other use cases
such as when the timeout is used as a backup for slow HWM convergence,
scaling is inappropriate.