modules/live: note lack of reduction scaling
The protocol for building the topology in the 'live' module
uses TIMEDFLUSH reduction.  Annotate the fact that the timeouts
are no longer scaled by TBON height, which will affect the performance
of this protocol.  It was not updated because 'live' is not
currently loaded by default, and this protocol may be reworked
in the near term.