cmd/flux-wreck: add --keep option to purge
Add a -k, --keep=CODE option to `flux wreck purge` to allow
arbitrary jobs to be skipped during purge. For example, the
following will keep all jobids that are multiples of 5:

 flux wreck purge -Rt 10 --keep='id % 5 == 0'

In the keep function, the only identifier currently exported
is `id` which maps to the currently process lwj id.