broker/hello: fix leak/error detection in flux_rpc
The flux_rpc() use in broker/hello.c seemed to assume integer return
and tested for error code < 0. Fix the call so that the flux_rpc_t
returned from the call is captured, and test for NULL on error.

This also allows the rpc object to be freed, fixing a small leak.