modules/libkvs: use fence to implement commit
Implement kvs_commit() as kvs_fence (name, 1), where name
is a uuid generated for the single commit.

This allows commits to complete in the same efficient way
as fences;  that is, piggybacked on the setroot event
resulting from the commit.

Add async interfaces: kvs_commit_begin() and kvs_commit_finish()