modules/kvs: special case "handler restart" wait_t
Implement a new, generic wait_create() that takes a callback
and arg.  When the wait_t usecount reaches zero, the callback
is called.

The "message handler restart" use case, formerly the only
supported use case, is now a special case.  To underscore this,
the interfaces were renamed:

  wait_compare_f       => wait_test_msg_f
  wait_create()        => wait_create_msg_handler ()
  wait_destroy_match() => wait_destroy_msg ()

Also, drop the time argument from wait_destroy().  This was
only used to time kvs_get(), which is not all that useful.
Drop "get" and other unused counters from KVS statistics.