modules/kvs: stall commit completion on store
Make the commit process "stallable" so that the KVS module
can remain responsive when content store requests are in progress.
Try to get all content store requests for a commit in flight
at once, asynchronously, to avoid boxcar effect on RPCs.

When a fence/commit becomes ready, it is added to the ctx->ready queue.
Only one commit can be active at a time.  When one finishes, or when
one is enqueued and afterwards, the queue contains exactly one entry,
begin processing the commit at the head of the queue.

Make commit_apply_fence() idempotent, and then arrange for it
to be stalled/restarted when
 * the rootdir or any directories required to walk the (changing)
   name space are not in local cache
 * any new objects are still dirty with respect to the content cache

Commit coalescing is still not (re-)implemented.