modules/kvs: send array of names in setroot event
Now that commits are uniquely named and completed via the setroot
event, instead of only one optional name in the setroot event,
send an array of them and require there to be at least one.

This prepares the way to reimplement commit aggregation.

Update proto test, and add diag output so encoded messages can be
scrutinized when the test is run manually.