modules/kvs: reimplement commit aggregation
Merge commits when multiple commits are in the ready queue
when the "check" watcher runs, and the current commit has not
progressed to the point of unrolling the rootcpy.

When a merged commit completes, the setroot event name field,
changed to an array in a previous commit, contains >1 name.
Log this when it happens on rank 0 at LOG_DEBUG level.

Add a module parameter "commit-merge=0" used to disable commit
merging for debug/performance testing.