modules/kvs: avoid storing root object redundantly
Problem: all ranks of the KVS, in response to a setroot event,
would synchronously store the root object to the content store
as a side effect of placing it in the local cache.

Instead of calling the generic store() function, manipulate
the local cache directly, inserting the root object without
triggering a flush to the content store.