modules/kvs: add commit/fence error path
Register a handler for the new kvs.error event, so that
fence requests can be finalized with an error code in the same
way that they were finalized with a success response by the
setroot event.

If the 'errnum' variable is set nonzero at any point during commit
handling on rank 0, the kvs.error event will be generated instead
of the kvs.setroot event, and the root blobref will not be updated.

Commits thus become "all or nothing".  If a commit/fence fails,
none of the operations contained within it were committed to
the namespace.