modules/kvs: add error handling to store path
Propagate errors from store(), the function that updates the local
cache and flushes to the content store, back to the commit handling
code, so errors can be returned to the caller of kvs_commit(), via
the kvs.error event.

Note: errors that occur in the content store continuation,
if invoked asynchronously, are *not* propagated.  This case
still needs to be handled.  The error reported in issue #776,
where the content store request failed with EBUSY because the
handle ran out of matchtags, would be handled, since matchtags
are allocated in flux_rpc() call.