modules/kvs: don't allow kvs put of key .
Catch "." as a special case during commit processing.
commit_link_dirent() now returns 0 on success, -1 on failure.
If this case fails, it will generate a kvs.error event
and the caller of kvs_commit() or kvs_fence() will get
an EINVAL error.

Add a test case to sharness t1000-kvs-basic.t for this.