bindings/lua: add flux.exitstatus decoder
Add a helper function to decode exit status integers to the flux
Lua module. Works like

 local state, code, coredump = flux.exitstatus (status)

where state is either "exited", "killed" or "stopped" accordding to
WIFEXITED(status), WIFSIGNALED(status), or WIFSTOPPED(status).

In the case of "exited" code is the exit code and coredump is nil.
In case of "killed" code is the terminating signal, aod coredump is
1 or true if a core was generated. In the case of "stopped", code
is the signal that stopped the process and coredupm is nil.