modules/kvs: add kvs_put_treeobj() function
Add kvs_put_treeobj() which links a key to a directory
entry object.  For now the only way to get a directory
entry object will be from kvs_get_treeobj(), to be
added in a later commit.

Add some cursory checks on the directory entry argument
to ensure incorrect metadata is added to the KVS.

Fixes #778