modules/kvs: don't merge commits to same key
Problem: a kvs.watch notification is not sent for every
update of a watched key if two or more commits operating
on the key are merged by the kvs for performance.  In
that case, a watcher is only notified once for the single,
merged commit.

This may be a problem for some use cases, for example
a finite state machine may use the kvs to communicate
state by commiting each change to a watched key.  Watchers
cannot follow the complete state if intermediate steps are lost.

Solution: check commits before merging them to see if they
change the same keys to different values, and abort the merge
if so.  This check does not apply to directories modified
as a side effect of a commit; only the fully qualified keys
supplied in the commit messages are checked for duplicates.