test/kvs: tighten up watch-mt test
The kvs watch-mt test starts N threads watching the same
key, then modifies the key monotonically M times.  Now that
watch is guaranteed to get a callback on every key change, this
test can be tightened up to verify monotonicity of the watched key.

Add a run of the test on rank 1 to ensure a "slave" KVS behaves
the same as the "master".