test/kvs: add dtree test
Add a program to create a H x W key namespace hierarchy.
That is, create W^H keys such that directories are H levels
deep and W entries wide.  The keys are created in one commit.

Use this in t1000-kvs-basic.t sharness test to create a
hierarchical key space and walk it with "flux kvs dir".