modules/job: avoid kvsdir_put() + cleanup
Instead of kvsdir_put() use kvs_put().

Drop extra kvs_commit() per new job that was required to allow
the kvsdir_t to be created.

Clean up add_jobinfo() to allow an error to be returned as was
originally intended, and to avoid creation of JSON object (and
malloc) in converting date to a JSON string for kvs_put().