modules/kvs: kvs.get codec: treeobj in/out
Allow a treeobj to be sent as an optional argument to a
kvs.get request, so that the key can be looked up relative
to a snapshot.

Echo this back in the response, or if it was not provided,
include a treeobj that identifies the current root that
was used for the lookup.

Update codec test

Adjust callers to compile with the new codec signature.
New functionality based on the codec change will be added in
a subsequent commit.