modules/kvs: handle treeobj args in kvs.get request
If a treeobj is provided in a kvs.get request, look up
the key relative to this "snapshot".

Return the treeobj used to look up the key in the response,
whether it be the one provided in the request, or one
representing the current root.

This may be useful in a future API that allows the current
version of a directory to be fetched and then walked as
a snapshot.