wreck: move wrexec.run service into the job module
Move the minimal code from the standalone `wrexec` module directly
into the `job` module. Do not load wrexec module in rc1 scripts
anymore since it now conflicts with job.

The code from wrexec was greatly simplified when moving it to the job
module. The weird parent-fd pipe is no longer used, and the double
fork before the exec of wrexecd was eliminated. For this case, these
features had no use, and a small speedup may be realized by removing
the extra code.

A `closeall` function was replaced with a use of `fdwalk` as well.