libflux/rpc: make flux_rpc_get() idempotent
Problem: flux_rpc_get() can only be called once on a
flux_rpc_t, which may be inconvenient or lead to poorly
structured code as discussed in pr #835.

In the single-response case, there is really no reason
to invalidate a received message while the flux_rpc_t is
valid.  The origin of this semantic is the multi-response
case, where a flux_rpc_get() would "consume" a message and
make way for the next one.

Replace flux_rpc_completed(), a simple test whether all
responses to an rpc have been received, with flux_rpc_next(),
which invalidates the current received message and returns 0
if another is expected, -1 if not.

Update flux-ping, flux-module, and flux-hwloc to reflect
this API change.

Update multrpc and kvs_watch sharness tests.