libflux/rpc: fix problem with lost errno
Problem: if flux_rpc_check() calls flux_recv() and gets an
error other than EWOULDBLOCK, a subsequent call to flux_rpc_get()
will ignore the fact that an error has occurred and retry the
flux_recv().

If an error occurs in the receive path, cache the errno in the
rpc object, and don't retry the receive, simply return the
cached errno.