modules/libjsc: update for hierarchical lwj directory
Remove all places where lwj.%d is assumed path/key to job id
in kvs.

When lwj path in kvs needs to be determined, look it up via
job.kvspath service and memoize result in a new ctx->job_kvs_paths
hash.

In job_state_cb, check for existence of "kvs_path" key in json
object, and if it exists, cache the kvs path for this jobid,
thus avoiding a job.kvspath query.