t/t0005-exec.t: Fix corner case in test for file not found
The fake file 'nosuchprocess' does not have a relative or absolute
path specified.  So the file will be searched for in all
directories in the user's PATH.  If the user does not have access
to one of the directories in their PATH, an EACCES error can occur
instead of ENOENT, leading to a failed test.  So put "./" in front
of 'nosuchprocess' to ensure it is not searched for in PATH.