modules/live: avoid using optparse for module opts
When experimenting with dlopen flags for modules, for
some reason t0010-generic-utils.t began failing due
to an option parsing error in the 'live' module.

live was using optparse to parse its arguments, but
optparse, which uses getopt(), is not MT-safe.  Switch
the module over to the simplistic key=val argument
parsing idiom used in other modules and update the test.