src/cmd/flux-ping.c: Fix corner case in ping output
If a rpc response contains an error and it is the last response
in a set, ping output will not be output as code follows an error
path in the last rpc.  Add ping_data struct to manage history of
ping responses and output ping info appropriately regardless the
order of rpc responses.