test: backtrace-all.sh: fix backtrace for exec in PATH
When executable causing backtrace is in PATH, and thus doesn't have
a relative or absolute path recorded in the corefile, the
backtrace-all script would fail to list a backtrace because
`test -x $exe` was failing. Precede the check with a `which $exe`
to translate an executable name to a path.

Also, quote around '*.core' in case there are one or more corefiles
in the current directory.