build: add script for generating release notes
Add script for generating a bulleted list of merge commits
since the last tag in github markdown.  Each entry consists
of a markdown link to the corresponding pull request and
the "body" description of the commit.  (The "subject" portion
is used up by the merge commit summary).  Multi-line bodies
are recognized so that each line remains part of the same bullet.