build: ensure jansson's json_int_t is 64 bits
jansson pack/unpack define the %I format token to correspond
to json_int_t, and json_int_t is either defined as long long
or long.  If the underlying system supports them, this will
always(?) be a 64 bit integer.

Rather than define a flux-specific type for json integers,
we make the guarantee in our API that %I is a 64 bit integer,
and catch it at build time if this likely assumption turns out
to be false.