Output appropriate error message on help failure
If flux-core is configured and built with --disable-docs, the "flux
help" command may still pick up manpages generated outside of the
current build (such as a version of flux installed on the system).

Install a simple ".nodocs" file when not generating documents.
When user attempts to user "flux help", this file can be detected
and an appropriate error message will be output to the user.

Fixes #874