libutil/blobref: create abstract blobref functions
Problem: SHA1 hash is hardwired in various parts of Flux,
but the "blobref" was designed in RFC 10 to accommodate
different hash functions.

Rename libutil/shastring to libutil/blobref, change the API
so that it conforms to the abstract design of RFC10, and
add support for sha256, mainly for testing.

Update users: kvs, content store, sqlite backing store,
and flux-content command line tool.