wreck: allow fatal kvs error without wrexecd exit
Allow a "fatalerror" to be written to kvs job "error log" without
actually causing wrexecd to immediately exit in wlog_fatal().

This is required for cases where we want to register and log a
fatal error, but allow wrexecd to continue startup protocol so
that barrier and kvs fence for other wrexecd processes can be
completed normally. The fatal kvs error will then be detected
and the job will terminate with a failed state instead of hang.