wreck: fatal error if nprocs <= 0
Register a fatal error in kvs if local number of processes to run
is detected as less than, or equal to, zero. This typically means
an "invalid specification" was inserted into the KVS for the lwj
in question, e.g. rank.X.cores == 0.

Do not, however, allow wrexecd to immediately terminate in
wlog_fatal(), instead allow all wrexecds to reach the "starting"
state transition kvs_fence(). After, wrexecd on rank 0 will notice
the fatal error in kvs and will abort the job with a "failed" state.

Fixes #900