libflux/rpc: Support special case strings
Support "any" and "upstream" nodeset strings in rpc_multi().
These will send a single RPC to FLUX_NODEID_ANY or
FLUX_NODEID_UPSTREAM respectively.  This support is for convenience,
so that any code that may need to send to FLUX_NODEID_ANY/UPSTREAM
is not required to setup a special case call to flux_rpc() instead
of flux_rpc_multi().