lru_cache: fix item loss on move-to-front
The issue eventually causes the LRU to count full but contain no
elements, when the NULL head or tail are used after that it becomes a
segfault.  It does not appear to be a simple pattern that happens every
time something moves, but only when a node with a null prev pointer is
not first, or a null next pointer in a node that is not last.