conn/local: switch to jannson interface
This resolves a free of data allocated by a different allocator detected
in jemalloc and tcmalloc from shortjson, always from this callsite.  The
issue does not appear to affect other uses of shortjson, or use with the
standard system allocator, despite the fact that this is a common
pattern and appears to match calls correctly.

In removing the last of the shortjson dependence, I noticed
that the sub and unsub functions were identical, so they now
also use a common helper function to reduce duplication.