lru_cache: reset lru list prev/next on removal
When a LRU entry was removed from the LRU list, failure to reset
the prev/next pointers to NULL on the removed entry could cause
an entry immediately re-inserted (as done in lru_requeue) to
corrupt the internal list. For example a move-to-front operation
would cause the front of list to have a prev pointer, which should
never happen.

Setting l->prev = l->next = NULL solves this problem as any insert
can be handled same as a new entry.