wreck/openmpi.lua: drop LD_LIBRARY_PATH, set orte_tmpdir
The new openmpi Flux component uses FLUX_PMI_LIBRARY_PATH
at runtime to dlopen our PMI library, so LD_LIBRARY_PATH
does not need to be set for it to work right in the
default configuration.  Therefore, don't change it
in the openmpi lua script.

Do set OMPI_MCA_orte_tmpdir_base to the rank-specific
directory so that shared memory segment names do not
collide when there are multiple Flux ranks on a (real)
node launching an MPI program with multiple MPI ranks
on a (flux) node.