test/generic-utils: run heaptrace subcommand
The flux heaptrace command and heaptrace service in the broker
had no test coverage.  We don't normally build --with-tcmalloc,
so the test just does start/dump/stop and accepts an exit 1
with "Function not implemented" on stderr as a valid result.
At least we will be exercising the RPCs and the error path in travis.