Make many .gitignore additions
There are many files that clog up the "git status" output at various
stages of the process.  We add them all to appropriate .gitignore files.

The biggest source of "git status" noise is from the test infrastructure.
Most tests that are auto-generated take the form test_*.t, so that ignore
is added at the top level.  The only conflict is with the test_runner.t
file, so an explicit exception is added for that.