test/coproc: avoid excessive VM usage
Problem: running 10,000 coprocs causes coproc_create()
uses 10000*2mb = 20gb of VM, resulting in ENOMEM on
a system with 32-bit pointers.

The requested amount of memory cannot be represented in
32-bit pointers.  In addition, the default vm.overcommit_memory
setting (0: heuristic overcommit handling) may reject an
allocation greatly in excess of physical memory.
Run 500 instead of 10K coprocesses, which is still
a reasonable test.

In addition, clean up the test so that it doesn't produce
a series of cascading segfaults and obscure failures when
this happens.  Fail the first test and continue on with
however many coprocs were able to be allocated.