modules/wreck: fix inconsitent \n use in wlog_fatal
Problem: wlog_fatal() sends log messages to flux_log() or
vfprintf (stderr, ...) depending on context.  The former
takes messages without newlines; the latter requires a newline.

Append a newline character to the stderr stream when that
mode is selected.  Drop trailing newlines from a couple
of wlog_fatal() calls to be consistent with the majority
of other calls.