local://: add userid/rolemask per RFC 12
The local:// connector authenticates users with SO_PEERCRED.
If the authenticated uid matches the effective uid of the broker
(the instance owner), no further lookup is performed and the
authenticated userid and rolemask of FLUX_ROLE_OWNER is cached
in the connection state.  If the authenticated userid is not the
instance owner, RFC 12 calls for a userdb lookup to determine
access and rolemask.  Since the userdb is not yet implemented,
the connection is dropped for non-instance owners.

Messages are created with userid field set to FLUX_USERID_UNKNOWN
and rolemask set to FLUX_ROLE_NONE.  Each message that is received
on a connection with userid and rolemask fields set to these
default values, or values matching those cached in the connection
state, are allowed.  Their userid and rolemask fields are rewritten
to the cached values before being forwarded to the broker.

Messages with other userid or rolemask settings are denied,
except as noted below.  If the denied message is a request, it
receives an EPERM response;  if it is another message type,
it is dropped.

As a special case for testing, if the rolemask cached in the
connection state includes FLUX_ROLE_OWNER, messages are allowed
to pass through with any userid/rolemask, unmodified.  This creates
a mechanism for testing service security policy without requiring
the ability to authenticate as a different user.  Note: this does
not grant the instance owner the ability to change its system
identity to the specified userid, since the instance does not have
any special system privileges.  Likewise, FLUX_ROLE_OWNER is a
superset of all other roles, so allowing FLUX_ROLE_OWNER to substitute
other roles can only decrease its capability.