shmem://: add userid/rolemask per RFC 12
The shmem:// connector is a point to point shared memory
connection between comms modules and the broker.
As such, connections are "authenticated" as the instance
owner and are given the FLUX_ROLE_OWNER role.

Any messages received with userid/rolemask set to
FLUX_USERID_UNKNOWN/0 are rewritten to the instance
owner/FLUX_ROLE_OWNER.  Other messages are allowed to
pass through so that connectors implemented as comms modules
can perform authentication and set the userid/rolemask
for other users.