loop://: add userid/rolemask per RFC 12
The loop:// connector is used in test and doesn't involve a real
broker connection.

Behave somewhat like the shmem:// connector.  If a message has purposely
set userid or rolemask, allow that setting to remain to facilitate testing.
If userid is FLUX_USERID_UNKNOWN, set it to the effective uid of the test.
If rolemask is zero, set it to FLUX_ROLE_OWNER.