test/rolemask: test message handlers in loop connector
The loop connector is a connector used only in testing.
It doesn't have connection state per se, so all messages
sent on it are given the default connection credentials
of geteuid()/FLUX_ROLE_OWNER.  This can be overridden
with flux_opt_set().

This test verifies that messages are rewritten as designed,
and that the default loop connector userid/rolemask can be
changed with flux_opt_set().

It then sets up a request handler and tests the handler policy,
adjusting both the policy and the connection credentials
through various permutations.