modules/userdb: add simple userdb module
The purpose of the userdb is to map userid's to a rolemask
for connectors, and to give the instance owner flexibility
in assigning roles to users.

When a user tries to establish a connection to the broker via
a connector, the connector authenticates the user and looks
up the user's rolemask in the userdb.  If a user is not in the
userdb, or has a rolemask of FLUX_ROLE_NONE, connection is
denied.  Otherwise, connection is allowed and the rolemask
and userid are stamped on messages sent from that connection.
These become input to multi-user security policy in each service.

This simple implementation is designed to be run only
on rank 0, since it has no mechanism to synchronize its
in-memory state on other ranks.  A future version might
use the KVS for shared state.

One entry for the instance owner is added when the module
is loaded.  This user has the FLUX_ROLE_OWNER role.
The instance owner is determined by calling geteuid().

If the --default-rolemask option is set at module load time,
users are automatically added to the userdb when they are
looked up, and are assigned the roles listed in the option's
arguments.  For example to give users a default role
of FLUX_ROLE_USER:

  flux module load userdb --default-rolemask=user

Methods are available to support a command line tool that can
manage the roles assigned to individual users after the userdb
module has loaded (e.g. in rc1 or later, interactively).