modules/connector-local: lookup users
After authenticating a user on the UNIX domain socket using
SO_PEERCRED, look the user up in the userdb to find what
roles have been assigned, if any.  If the userdb cannot find
the user, or the user is assigned no roles, drop the connection.

Since the local connector has to be functional before modules
such as the userdb module are loaded by rc1, the userdb is
bypassed if the authenticating userid matches geteuid().
In this case, FLUX_ROLE_OWNER is assigned without a userdb lookup.