cmd/flux-event: fix unsubscribe logic
Problem: flux-event --count unsubscribes from events when it exits
its event receive loop.  If topic strings are specified on the
command line, the command line is dereferenced incorrectly.

Use the same logic to dereference the command line for unsubscribes
as was used for subscribes.

Note: this problem was previously masked by the lack of error
propagation in the local connector's unsubscribe method.