libpmi: Change ops from pass-by-reference to pass-by-value
Change the 'ops' parameter of pmi_simple_server_create() from
pass-by-reference to pass-by-value.  Previously the function made
an internal copy of the ops structure, freeing the caller from
having to maintain the ops structure for the life of the
pmi simple server.  However, the only way to know that was to read the
code.

By changing to pass-by-value it is immediately clear to the caller,
even without a comment, that the pmi simple server code can have no
reference to the caller's struct.