flux-start: Add --bootstrap method.
Add the --bootstrap option to allow specifying a specific instance
bootstrap method.  The initial two methods are "selfpmi" and "pmi",
which are just explicit names for existing methods.  selfpmi is the
new name for the method where flux-start itself starts a PMI server,
and then launches brokers use PMI to bootstrap using said server.
With "pmi", flux-start just assumes a PMI environment is available
and execs a single broker directly.

Documentation is updated and tests are added for this new interface.

Fixes #952