libflux/response: reset userid/rolemask on reply
When deriving a response message from the request,
clear the userid and rolemask so they will be set
by the connnector.

If this is not done, then on connections authenticated
as the instance owner, responses may be tagged with
the sender's userid/rolemask, which may be rejected
by the reactor's message dispatch engine.