broker/attr: allow any role to get and list broker attrs
Allow all broker attrs to be fetched (but not set) by any user.
Only the instance owner can set broker attrs.

This is mainly to allow size and rank to be fetched.
In the future the security policy could be extended
to allow each attribute to have mode bits if we want.