connector/local: add single byte auth response
Problem: local connector allows connect to succeed, then
drops the connection abrubtly when authentication fails.
On the client end, flux_open() succeeds but the next
operation such as a flux_send() will fail with EPIPE.
This error is not particularly helpful for the user.

Add a single byte authentication result to the local wire
protocol.  Return 0 on success or errno on auth failure,
so that on failure, the client flux_open() will get EPERM.