broker: add userid, rolemask to ping response
Consolidate "ping handlers" in broker and modservice, then
add userid and rolemask JSON fields in the response so that
we can see what userid and rolemask arrived in the ping
request.

Update flux-ping to parse the new JSON fields, and
add --userid option to display them in the ping output.
Update flux-ping(1) manual page for new option.