broker/log: add optional reply to log.append RPC
Problem: with new multi-user security features, a log request
can be denied by the message dispatcher if it comes from a user
other than the instance owner.  In some case flux_log() users
may want to treat this as a fatal error, but they can't if
there is no RPC response.

If the log.append request was sent with FLUX_MATCTHAG_NONE, there is
no response as before.  Otherwise, a success/fail response is returned
to the sender.