libflux/flog: add FLUX_LOG_CHECK flag
Problem: flux_log() cannot return an error.

If the FLUX_LOG_CHECK bit is set in the severity parameter
passed to flux_log() or flux_vlog(), make the log.append RPC
with flags=0 and block until a response is received.  Otherwise,
make it with flags=FLUX_RPC_NORESPONSE and return immediately,
as before.

Change the return value of these functions from void to int.
If the FLUX_LOG_CHECK bit is set, a return of -1 with errno set
is possible.  Otherwise, the return value will always be 0.