broker: install SIGALRM timeout for teardown
Problem: at times shutdown hangs

The code for unloading modules and closing zeromq sockets
needs work, but in the mean time, set an alarm clock and
exit if post-reactor cleanup takes longer than 1s.  This
period is after the hueristically-determined shutdown
period in which the instance can still communicate because
the reactor is still running.  Only local stuff can happen
after the reactor has terminated, but some modules may be
still be waiting for messages that will never arrive if
they failed to unload during the shutdown period.

Preserve the exit code if the alarm clock fires.  E.g. timeout
is not handled as an error at this time.  Before we used to
just exit without teardown when the shutdown period was over,
so this is certainly no worse than before.