broker/snoop: drop snoop facility
Problem: the "snoop" facility uses the deprecated czmq
zsocket class.

Other difficulties: the snoop faciliity no longer captures
all the messages passing through the broker as its hooks
have not been kept up to date.  It uses a seperate socket
when it probably should be built as a regular flux service.
The socket is created on demand, but listing the snoop-uri
attribute (including lsattr -v) causes it to be created,
which is an open issue.

Drop the snoop facility for now, including broker class,
broker hooks, command, tests, and man page.

Fixes #455