broker/module: address broker hangs in pthread_join
Problem: broker teardown sometimes gets stuck in
pthread_join() on modules that don't complete
unloading in rc3.

Modules that are still loaded when the shutdown grace
timer expires are unloaded in the post-reactor teardown
code.  First a shutdown message is sent to each module,
then the modhash is destroyed.  During modhash destruction,
the hash element destructor is called on each remaining
module.  The destructor calls pthread_join() on the module
thread before attempting to free its resources.  This may
hang, for example if the module thread is blocked on an
RPC with the broker reactor stopped, and the module hasn't
yet processed the shutdown request.

Change the modhash destructor so that, as a prelude to
destroying the hash, it iterates over the remaining modules
in the hash, calling pthread_cancel() on them.  This
makes it much more likely that the pthread_join() will
be successful.

Since this means some teardown within the module thread
may have been skipped, log each module that is terminated
with a cancellation so we can run those down.

Drop dead code for stopping/starting all modules asynchronously.